Modified the metamodel and the model building in order to export also the number of core and the processing rate for each container and the global speed norm factor.

+++ src/main/java/it/polimi/modaclouds/adaptationDesignTime4Cloud/Main/MonitoringRulesHelper.java
@@ -39,7 +39,7 @@ public class MonitoringRulesHelper {
 					
 					if(t.getFunctionality().size()>0){
 					
-						float threshold= t.getResponseTimeThreshold().get(0).getValue();
+						double threshold= t.getResponseTimeThreshold().get(0).getValue();
 						
 						rule= factory.createMonitoringRule();
 						rule.setMonitoredTargets(factory.createMonitoredTargets());
